pyproject.toml
[project]
name = "airflow-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "airflow-operators>=0.11.0",
    "apache-airflow>=3.2.0",
]
